python/requirements.txt
nltk>=3.7
numpy>=1.21.0
python-docx>=0.8.11

